Move sheet reader code to the WorksheetPartReader
